Refactor camelCased files to snake-case.
Mac OSX is case insensitive, but Heroku is case sensitive. models/foodtruck was renamed to models/foodTruck, but that change was not reflected in the git history. So then Heroku was unable to find the module: http://stackoverflow.com/questions/19341975/heroku-node-cannot-find-module-error

Snake casing follows the module-naming conventions of other projects like express, and avoids these types of errors in the future.