Start refactoring mongoose -> MongoDB. Food truck operations in mongodb-test.
Mongoose has been difficult to debug, because it's an abstraction of MongoDB. Also, there are certain operations that are more efficient in MongoDB (collection.insert() vs Model.create(), for example). Also, not all of the Mongoose methods are promisified.

Given the above points, and because I'm realizing I don't have as solid of a grasp on MongoDB as I'd like, I'm refactoring to use MongoDB, specifically promised-mongo.