Huge refactor of database logic to use MongoDB in routes and worker.
Now promised-mongo is used everywhere for database logic. The worker task, run by heroku, now works. Before, with mongoose and mongoomise, the food truck documents were not being inserted into the food trucks collection.

I wasn't able to abstract the raw promised-mongo queries into model functions, so there's a bunch of database business logic in the routes and the worker.

Another problem is that there's no error handling in the promises. The promises appear to only return one value. So I need to figure out how to tell whether or not the promise has returned an error, and handle the error.