Stop tracking the dev version of bin/getAPIData to keep history cleaner.
`src/workers/get-api-data.js` will remain part of the repo and be used for testing the worker task, but it has nearly the same content as bin/getAPIData, so it should not be commited. The file was added to .gitignore in the last commit.

Ideally I'd have some automatic task that modifies get-api-data.js (the hashbang header) and regenerates bin/getAPIData every time get-api-data is modified, but I haven't looked into how to do that. So for now my workflow is:

- run `node src/workers/get-api-data.js` until the script works. Copy all of the code except for the hashbang header to bin/getAPIData. Commit the changes to bin/getAPIData.